+++ .eslintrc.js
@@ -13,5 +13,6 @@ module.exports = {
   plugins: [
   ],
   // add your custom rules here
-  rules: {}
+  rules: {
+  }
 }
