+++ components/protected/layout/app_ber/AppBarTab.vue
@@ -1,0 +1,22 @@
+<template>
+  <v-tabs align-with-title>
+    <v-tab>Tab 1</v-tab>
+    <v-tab>Tab 2</v-tab>
+    <v-tab>Tab 3</v-tab>
+  </v-tabs>
+</template>
+
+<script type="ts">
+  import { defineComponent } from '@nuxtjs/composition-api'
+
+  export default defineComponent({
+    setup(){
+      
+
+      return {
+
+      }
+    }
+  })
+
+</script>
