+++ middleware/authenticated.ts
@@ -1,9 +1,13 @@
-import { defineNuxtMiddleware } from "@nuxtjs/composition-api";
+import { defineNuxtMiddleware } from '@nuxtjs/composition-api'
+import { UserStore } from '~/store'
 
-// 未ログインの場合ログインページに飛ばす
-export default defineNuxtMiddleware(ctx => {
-  if (!ctx.store.getters['user/isAuthenticated']) {
-    const snackbarPayload = JSON.stringify({ color: 'warning', message: 'ログインしてください' })
+// 未ログインの場合ログインページに飛ばすs
+export default defineNuxtMiddleware((ctx) => {
+  if (!UserStore.isAuthenticated) {
+    const snackbarPayload = JSON.stringify({
+      color: 'warning',
+      message: 'ログインしてください',
+    })
     const encodedPayload = encodeURI(snackbarPayload)
 
     ctx.res.setHeader('Set-Cookie', `snackbar=${encodedPayload};Path=/`)
