+++ nuxt.config.js
@@ -21,9 +21,8 @@ export default {
   // Plugins to run before rendering page: https://go.nuxtjs.dev/config-plugins
   plugins: [
     '~/plugins/axios.ts',
-    '~/plugins/firebase.ts',
     '~/plugins/vee-validate.js',
-    '~/plugins/auth.client.ts'
+    '~/plugins/auth.client.ts',
   ],
 
   // Auto import components: https://go.nuxtjs.dev/config-components
