+++ tsconfig.json
@@ -27,7 +27,7 @@
     "types": [
       "@nuxt/types",
       "@nuxtjs/axios",
-      "@types/node"
+      "@types/node",
     ]
   },
   "exclude": [
